Apply Alexandre Rosenfeld's DACP patch from Google Summer of Code 2010
Add DACP support. This allows an application such as Rhythmbox to be
controlled using a remote control application on an iPhone, iPod Touch,
iPad or other similar device. The DACP code is not yet complete, but
progress warrants commiting it to the development tree and focusing
future work on specific patches.
Signed-off-by: W. Michael Petullo <mike@flyn.org>