Re-license some files as LGPL
I had accidentally placed GPL headers on some of the source files for
which I maintain the copyright. I have replaced these with LGPL headers.
Alexandre Rosenfeld mistakenly (I assume) licensed dacp-player.[ch] as
GPL. I will have to confirm that it is okay to change these with him.
Andre Moreira Magalhaes' test-dmap-client and test-dmap-server are also
GPL. These files are not as much of an issue, but I plan to contact him
and ask if LGPL is okay so that all source files are consistent
Signed-off-by: W. Michael Petullo <mike@flyn.org>