dmap-share: Don't throw warnings during normal operations
DACP (_touch-able._tcp) and DAAP (_daap._tcp) on macOS use the same
default port, 3689, to advertise both services. A single server likely
implements both protocols, and internally routes the requests.

This is not the case with the libdmapsharing, with each protocol getting
its own server running on the port (both in IPv4 and IPv6). When an
application like rhythmbox would launch both possible types of share,
the first one would go without a hitch, the second one would only be
accessible via IPv6 on a different port.

This doesn't make much difference as:
- which port is used is ultimately unimportant as the port is
  advertised in mDNS
- the port already being used might mean we're sharing music with
  another application on the same machine
and none of those are programming or configuration errors.

Ideally, services with the same port would share a single SoupServer,
but until then, this change will not make any difference in terms of
interoperability.

https://bugzilla.gnome.org/show_bug.cgi?id=766723

Signed-off-by: W. Michael Petullo <mike@flyn.org>