Adjust how building tests works
Previously, the presence of the check library led to libdmapsharing
building its unit tests and supporting facilities, despite the separate
configure flag --disable-tests. This led to confusion as to whether some
test facilities should be present in the compiled libdmapsharing. Now,
--disable-tests disables all tests, whether unit tests or sample programs,
and it ensure not test facilities exist in libdmapsharing.