build: Fix vala build with gstapp in non-standard location
Some of the headers include `<gst/app/gstappsink.h>`, which is part of gstreamer-app-1.0,
the pkg-config file only lists gstreamer-1.0 as a dependency, though. It is not guaranteed
that they share the installation paths and indeed it is not the case on NixOS, resulting
in build failure.

Let's properly declare our dependencies so that vala-gen-introspect can find them.

https://gitlab.gnome.org/GNOME/libdmapsharing/issues/11